bundle in the licenses of poke's dependencies
Resolves #3

This is an attempt to comply with the licensing requirements of all of
the libraries that poke transitively depends on. I am not a lawyer. I am
not aware of a standard way to do this for CLI tools like this (it
really seems like that should exist!). This is my best guess at how to
comply.