bump geth version to enable cross-platform builds
Geth uses CGO, but as of geth commit a184ab7a61884409a4deea2419c5307a0dc7c49d,
the libraries that poke uses can be compiled with CGO disabled. This
enables cross-platform builds, eg:

    $ GOOS=windows CGO_ENABLED=0 go build
    $ GOOS=linux CGO_ENABLED=0 go build
    $ GOOS=darwin CGO_ENABLED=0 go build

This change also required a minor change to one of our library calls,
due to a backward-incompatible change in geth. I haven't tested this
with a real hardware wallet, but it should be fine.